# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Delivery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all food categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # RESTAURANTS
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List restaurants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantPage"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDetail"
    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    delete:
      operationId: deleteRestaurant
      tags: [restaurants]
      summary: Delete a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (for current user role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENTS
  /payments/create-intent:
    post:
      operationId: createPaymentIntent
      tags: [payments]
      summary: Create a Stripe payment intent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntent"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment and update order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmInput"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a restaurant
      parameters:
        - name: restaurantId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # NOTIFICATIONS
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    # RESTAURANTS
    Restaurant:
      type: object
      required: [id, ownerId, name, address, status]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        avgRating:
          type: ["number", "null"]
        totalReviews:
          type: ["integer", "null"]
    RestaurantDetail:
      type: object
      required: [id, ownerId, name, address, status, menuItems]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        avgRating:
          type: ["number", "null"]
        totalReviews:
          type: ["integer", "null"]
        menuItems:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
    RestaurantPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Restaurant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RestaurantInput:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        image:
          type: string
    RestaurantUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]

    # MENU
    MenuItem:
      type: object
      required: [id, restaurantId, name, price, available]
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        available:
          type: boolean
...
    Order:
      type: object
      required: [id, userId, restaurantId, totalAmount, status, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, payment_completed, accepted, preparing, ready_for_pickup, out_for_delivery, delivered, cancelled]
        deliveryPartnerId:
          type: ["integer", "null"]
        deliveryPartnerName:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, menuItemId, name, quantity, price]
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        image:
          type: ["string", "null"]
    OrderPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrderInput:
      type: object
      required: [restaurantId]
      properties:
        restaurantId:
          type: integer
        deliveryAddress:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, payment_completed, accepted, preparing, ready_for_pickup, out_for_delivery, delivered, cancelled]
        deliveryPartnerId:
          type: ["integer", "null"]

    # PAYMENTS
    PaymentIntentInput:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: integer
    PaymentIntent:
      type: object
      required: [clientSecret, amount, currency]
      properties:
        clientSecret:
          type: string
        amount:
          type: number
        currency:
          type: string
    PaymentConfirmInput:
      type: object
      required: [orderId, paymentIntentId]
      properties:
        orderId:
          type: integer
        paymentIntentId:
          type: string
    Payment:
      type: object
      required: [id, orderId, amount, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        amount:
          type: number
        transactionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        createdAt:
          type: string

    # REVIEWS
    Review:
      type: object
      required: [id, userId, restaurantId, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        restaurantId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [restaurantId, rating]
      properties:
        restaurantId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]

    # NOTIFICATIONS
...