# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Delivery Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  restaurantId: number;
  restaurantName?: string;
  totalAmount: number;
  status: OrderStatus;
  /** @nullable */
  deliveryPartnerId?: number | null;
  /** @nullable */
  deliveryPartnerName?: string | null;
  createdAt: string;
  items: OrderItem[];
}