# lib/api-zod/src/generated/types/paymentIntentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Delivery Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentIntentInput {
  orderId: number;
}